Upgrade Pex to 2.1.135. (#18876)
This pulls in support for Pip 23.1.1 & 23.1.2 and the new
`pex3 venv {inspect,create}` commands.

Pants likely has use for the `pex3 venv create` command in an overhaul
of AWS Lambda function and GCF support as well as in possible new
direct support for AWS Lambda layers.

The change-log is here:
  https://github.com/pantsbuild/pex/releases/tag/v2.1.135