Do not choke on missing lockfiles.
Fixes #18933 (and #18404). Original attempt at fixing this but failed
was in #18406 due to an oversight that it was the `_lockfiles` target
generator with a `sources` field that was being synthesized, not the
`_lockfile` target with a single `source` file only.

Either use should work equally well, and now they do.