Ensure packaged artefacts are fully replaced in dist/ (#18930)
This patch has the package goal clear out the artefact(s) it is about to
write to `dist/`, if they already exists.

For instance, if running ``pants package path/to:target`` on a
`pex_binary` (outputting `path.to/target.pex`), pants will now first
remove anything that's already at `dist/path.to/target.pex` before
writing the new output.

This resolves two problems:

- if there's existing contents of a different kind (e.g. a directory in
`dist/` and writing a file), the package call would explode. For
instance, switching a target like `pex_binary(..., format="zipapp")`
(file) to `pex_binary(..., format="packed")` (directory).

- if the package output is directory, stale files already in that
location in `dist/` would remain. For instance, a `pex_binary(...,
format="packed")` where a file was removed.

This fixes #17758 and fixes #18849, respectively.

This only fixes `package`, not any other goals that also write to fixed
paths (like `export` and `export-codegen`). In #18871, I start on
`export-codegen`, but it's a bit fiddlier (requires propagating "this is
the artefact" paths around) and it's best to land the infrastructure in
this PR first. I'll file follow-up issues covering them specifically.