graph: first pass attempt at garbage collection
Pushing this now to know if this is directionally or generally correct.

This is an attempt at solving #7675.  I'm aware there needs to be tests
on the graphs, but I'm still trying to load the data structure semantics
into my head to know how to construct a test.  In the meanwhile, I'd like
some feedback to see if this is going in the right direction or if I'm
totally off base here.

Also feel free to correct me or push me to better practices on my Rust.