Support Python requirement target addrs in tool requirements. (Cherry-pick of #19014) (#19023)
Currently users can enumerate `requirements` for a Python tool, which
will cause
Pants to use that requirement subset from the tool's lockfile. This
supports resolving
a tool from a larger lockfile, when that makes sense. If no requirements
are provided
the entire lockfile is used.

Two examples of when resolving a tool from a subset of a larger lockfile
is useful:
1. Resolving `pytest` as a tool at the same version as its runtime
library without
  specifying it in two places.
2. Having a single lockfile for all tools, and user code, to be consumed
by VSCode.

However, the value of this `requirements` option is typically redundant
with the
input requirements used to generate the lockfile. 

This change allows `requirements` to specify addresses (or address
specs).
This can be used to avoid the redundancy: You can point to some
python_requirement targets that are a subset of all those used to
generate
the lockfile.

The implementation turns out to be reasonably neat, as it just pushes
some code that turns addresses into requirement strings to slightly 
later in the pex creation logic, and actually allowed us to reduce some
redundancy and trim some code.

This addresses #18816.