Only build and release wheels for Python 3.9. (#18829)
Now that scie-pants is the norm, and it provides Python 3.9,
there is no need to release for 3.7 or 3.8. 

This will save us CI time as well as PyPI space.

This is a step on the road to deprecating running pants
in any way other than scie-pants. For now we still
have the escape hatch of installing Python 3.9 and
running from a `./pants` script.

Once we're comfortable with this, a future change can 
clean up the CI scripts, ensure that tests always run on
3.9, etc. For now this is the smallest change that 
achieves the desired effect.