Bring the tool `--export` option deprecation forward.
This is largely moot - in 2.18.0dev0 we're requiring you
to specify the resolves you want to export explicitly
via the --resolve option, so this --export option will
become a no-op. Still, it's good to deprecate it in case
anyone has it set in a pants.toml.

The later deprecation wasn't in any stable release yet, so
it's OK to change it.