Bring the tool `--export` option deprecation forward. (#19079)
This is largely moot: in 2.18.0.dev0 we're requiring the user to specify
the resolves they want to export explicitly via the --resolve option, so
this --export option will be a no-op. 

Still, it's good to deprecate it now so people can remove it from their
config files and know not to expect it to have any effect.

This deprecation wasn't in any stable release yet, so it's OK to change
it.