Run `terraform init` as part of `check` goal (#18974)
Many Terraform tools require Providers and Modules to be installed with
`terraform init` for them to work properly. This includes `terraform
validate` which is the backbone of the current `check` goal for
Terraform.
This is the simplest way of installing these. It makes no effort to
actually understand what modules are used, and defers entirely to the
`terraform init`. As the providers will be downloaded as one block, this
might take up a lot of space in the pants cache (I'm not sure tho).
We could do smarter things. Terraform has separate was of listing the
required providers; saving lockfiles of them; and downloading the
providers. (We could also parse the version specs from the Terraform
sources themselves and make lockfiles ourselves.) That sounds like a
cool thing to do, but definitely more complicated than bundling the
dependencies into a Digest. I think it's worth getting this support
working since `check` is not functional for any Terraform sources which
use Providers or Modules, which is probably most of them. This will also
allow us to integrate more Terraform tools. I think the improvements to
Provider fetching are unlikely to resolve real problems. Terraform
deployments usually pull in all their sibling files, contrasting with
Python where siblings might not be imported; therefore, we are unlikely
to see large gains in separating the downloaded Providers.

resolves #18489