Simplify interpreter-related CI config. (#19096)
Now that we only publish wheels for Python 3.9 we can trim away the 
affordances for multiple Python versions in CI.

We also pin the IC for Pants itself, in pants.toml and the relevant
dists.

We also regenerate our custom tool lockfiles for just the IC we care
about. Note that are not the built-in lockfiles we ship to users, but the
internal custom ones.

This dragged in a flake8 upgrade that complained about a few unnecessary
dict comprehensions, so this fixes those.

Note that this gets rid of the cron-based extended Python testing
workflow, since all that did was run the same tests on other interpreters,
which is now moot. Plus, no one ever looked at failures in those jobs anyway.