Warn about `files()` in PEXes in one place (Cherry-pick of #19027) (#19148)
There's several places that check whether a PEX transitively depends on
any `files()` targets, because those are (currently) not included in the
PEX. This moves the warning into the "pex from targets" builder, to not
have to reimplement it everywhere.

This is follow up to #19022, and restores the warning for
`layout="zip"`, since I didn't implement it there.

Co-authored-by: Huon Wilson <huon@exoflare.io>