Check for `CI=true`, not `CI=1` for cherry-picking (#19141)
This fixes #19127, which accidentally checked for `CI=1` to determine if
it's running on CI, but CI services (and GHA in particular) actually
[set
`CI=true`](https://docs.github.com/en/actions/learn-github-actions/variables#default-environment-variables),
and so the script was operating in (unchanged) local mode on CI 💥

This will fix a bug with #19125 if running on the second commit on
`main` (i.e. `HEAD^`), and so adjusts the `fetch-depth` comment to
explain that. This patch also ensures the shallow-clone optimisations
apply on CI.