Avoid extra `.` in `from .. import` parent imports in Rust Python dep inference (#19175)
This fixes #19173 by generalising the `from . import a` special case to
also handle `from .. import a` etc.: _any_ module path that ends with a
`.` doesn't need an extra `.` to separate the module path from the
name(s).

Before this fix, `from . import a` was special cased to avoid an extra
dot (becoming `.a` as desired), while `from .. import a` was not special
cased. It became `...a` (extra `.`), by joining the module path (`..`)
and the name (`a`) with the usual `.` separator. This separator is only
appropriate for normal names like `from x import a`.