Include committer date in local version identifier of unstable builds (#19179)
Fixes #7399 by adjusting the local version identifier of 'unstable'
wheels to also include a timestamp, so that builds with the same stable
version can be sorted in a vaguely sensible way, without having to
dereference the git hash manually.

The timestamp chosen is the committer date of the commit in question,
formatted as a vaguely human readable number (`yyyymmddHHMM`). For
example, this PR's CI built
`pantsbuild.pants-2.18.0.dev0+202305302357.git111f44db-cp39-cp39-macosx_11_0_arm64.whl`,
corresponding to 2023-05-30 23:57.