Rename `python_awslambda` to `python_aws_lambda_function` (#19216)
This renames the `python_awslambda` target to
`python_aws_lambda_function`. This is to reduce ambiguity now that Pants
has support for building layers via `python_aws_lambda_layer` (#19123),
in addition to functions, and also aligns with the
`python_google_cloud_function` target too.

This change has a single release deprecation since it's easy to fix, and
keeps it inline with the layout switch over (e.g. #19122), but it's also
easy to keep around for longer.

This PR essentially just does a search and replace on the documentation,
it doesn't update the docs to include info about layers. I'm working on
that separately.