internal: Create dep inference request type (#19001)
Boilerplate to introduce a request type for native dependency inference
suggested here:
https://github.com/pantsbuild/pants/pull/18985#issuecomment-1546643872

Adds a "metadata" field to the CacheKey type. Is it naive to store the
string as is in the key? Afaict, the `PersistentCache` is addressed by
the hash of the bytes of the key, not the key bytes themselves -> Should
be fine.

---------

Co-authored-by: Joshua Cannon <joshdcannon@gmail.com>