Fix secondary ownership warning semantics (Cherry-pick of #19191) (#19224)
Fixes #19174 by boiling the criteria to warn down to:

1. The Address/Target/FieldSet is matched due __only__ to
`SecondaryOwnerMixin`
2. It was matched due to a file spec (literal or glob)
3. It wasn't matched due ton address spec
- This is just to handle the corner case where it was matched due to
file and address

In order to facilitate 1., `Owner` and `find_owner` have been
(temporarily) modified. Additionally, the rule that goes from
`RawSpecsWithOnlyFileOwners` -> `Addresses` has been split to return
`Owners`, with an additional rule added to complete the graph (`Owners`
-> `Addresses`).

Then 2 and 3 are facilitated by just requesting the addresses based on
the include specs.