store stdout/stderr output from tests as bytes (#19184)
## Problem

Pants expects test output to be in UTF-8 even though this assumption is not a valid to make given that test output comes from process invocations. As described in https://github.com/pantsbuild/pants/issues/19121, this causes a `UnicodeDecodeError` under the conditions mentioned there.

## Solution

Instead of storing a `str`, just store the binary output for stdout and stderr from the underlying `Process` to avoid string encoding issues as much as possible. When creating a message for output, only then decode the stings as UTF-8 and replace invalid characters.

The `stdout` and `stderr` attributes are renamed to `stdout_bytes` and `stderr_bytes`, respectively, Introduce `stdout` and `stderr` properties so that external plugins get a deprecation warning.

Fixes https://github.com/pantsbuild/pants/issues/19121.