Turn ShouldTraverseDepsPredicate into a dataclass
closure functions have a useless hash. The hash of a function does not
account for any closure vars, including unhashable mutable vars,
so that hash is not reliable.
Plus, every instance of a closure will have a different id, so there cannot be
any reuse of anything cached based on the function's hash.

Using a frozen dataclass bypasses these issues. We include the class's __call__ function
as a field in the parent dataclass so that it is part of
the __hash__ function generated by the
@dataclass decorator. Using the class's __call__ is more stable than using an instance's function.