Replace `include_special_cased_deps` flag with `should_traverse_deps_predicate` (#19272)
This PR is extracted from #19155 which provides a way to stop dependency
traversal through package targets. Several other improvements will
probably benefit from this predicate as well (including proposals around
multiple dependencies fields and/or tagged dependencies).

### Passing a predicate on `Request` types

This PR replaces the `include_special_cased_deps` flag (on
`DependenciesRequest`, `TransitiveTargetsRequest`, and
`CoarsenedTargetsRequest`) with the `should_traverse_deps_predicate`
predicate:

https://github.com/pantsbuild/pants/blob/b85612740209309b40eee5f7a653f9ff163f0afa/src/python/pants/engine/target.py#L2519-L2521

https://github.com/pantsbuild/pants/blob/b85612740209309b40eee5f7a653f9ff163f0afa/src/python/pants/engine/target.py#L919-L927

https://github.com/pantsbuild/pants/blob/b85612740209309b40eee5f7a653f9ff163f0afa/src/python/pants/engine/target.py#L882-L887

Here is the `ShouldTraverseDepsPredicate` base class. This is a
dataclass that serves as the "closure" around the `__call__` predicate
function; if it needs any data, the subclass should add fields for that
data:

https://github.com/pantsbuild/pants/blob/b85612740209309b40eee5f7a653f9ff163f0afa/src/python/pants/engine/target.py#L717-L739

### Calling the requested predicate

The `resolve_dependencies` rule is responsible for calling this
predicate in two places.

1. To bail out early and avoid dep hydration and inference if unneeded:

https://github.com/pantsbuild/pants/blob/b85612740209309b40eee5f7a653f9ff163f0afa/src/python/pants/engine/internals/graph.py#L1320-L1326

2. To replace the `include_special_cased_deps` flag on a per-field
basis. If there are no `SpecialCasedDependencies` fields, then the
predicate won't be called.

https://github.com/pantsbuild/pants/blob/b85612740209309b40eee5f7a653f9ff163f0afa/src/python/pants/engine/internals/graph.py#L1411-L1416

### Replacing the `include_special_cased_deps` with the predicate

This predicate replaces `include_special_cased_deps` to minimize the API
around dep calculation as @stuhood recommended in
https://github.com/pantsbuild/pants/pull/19155#discussion_r1213566914:
> IMO, a true benchmark for this filtering API would be that we can
replace the `include_special_cased_deps` flag with it.
>
> I am very much in favor of adding an API to allow for filtering like
this, but I think that from a complexity-budget perspective,
killing/replacing "special-cased-deps" would be a really, really good
thing to accomplish here.

The default predicate, `TraverseIfDependenciesField` uses
`isinstance(field, Dependencies)` which is `False` for any
`SpecialCasedDependencies` subclasses:

https://github.com/pantsbuild/pants/blob/b85612740209309b40eee5f7a653f9ff163f0afa/src/python/pants/engine/target.py#L742-L750

Any thing that used `include_special_cased_deps=True` (the default was
`False`) now passes in the `AlwaysTraverseDeps` which always returns
`True`.

https://github.com/pantsbuild/pants/blob/b85612740209309b40eee5f7a653f9ff163f0afa/src/python/pants/engine/target.py#L753-L760

These are examples of passing in that predicate:

https://github.com/pantsbuild/pants/blob/b85612740209309b40eee5f7a653f9ff163f0afa/src/python/pants/backend/project_info/paths.py#L117-L130

---------

Co-authored-by: Joshua Cannon <joshdcannon@gmail.com>