Write directly to relevant notes file in changelog.py (#19303)
This updates `changelog.py` to write the (non-internal) changes to the
relevant release notes file directly, rather than requiring a human to
copy-paste it in. For now, the file is just mutated, without committing.
The internal changes are still printed for the human to deal with.

For instance, `pants run src/python/pants_release/changelog.py --
--prior 2.18.0.dev1 --new 2.18.0.dev2` adds a new section to
`src/python/pants/notes/2.18.x.md`:

```diff
diff --git a/src/python/pants/notes/2.18.x.md b/src/python/pants/notes/2.18.x.md
index e648a4525c..d6668a24b1 100644
--- a/src/python/pants/notes/2.18.x.md
+++ b/src/python/pants/notes/2.18.x.md
@@ -1,5 +1,35 @@
 # 2.18.x Release Series
 
+## 2.18.0.dev2 (Jun 14, 2023)
+
+### New Features
+
+* Include complete platforms for FaaS environments for more reliable building ([#19253](https://github.com/pantsbuild/pants/pull/19253))
+
+* Add experimental support for Rustfmt ([#18842](https://github.com/pantsbuild/pants/pull/18842))
+
+* Helm deployment chart field ([#19234](https://github.com/pantsbuild/pants/pull/19234))
+
+### Plugin API Changes
+
+* Replace `include_special_cased_deps` flag with `should_traverse_deps_predicate` ([#19272](https://github.com/pantsbuild/pants/pull/19272))
+
+### Bug Fixes
+
+* Raise an error if isort can't read a config file ([#19294](https://github.com/pantsbuild/pants/pull/19294))
+
+* Improve handling of additional files in Helm unit tests ([#19263](https://github.com/pantsbuild/pants/pull/19263))
+
+* Add taplo to the release ([#19258](https://github.com/pantsbuild/pants/pull/19258))
+
+* Handle `from foo import *` wildcard imports in Rust dep inference parser ([#19249](https://github.com/pantsbuild/pants/pull/19249))
+
+* Support usage of `scala_artifact` addresses in `scalac_plugin` targets ([#19205](https://github.com/pantsbuild/pants/pull/19205))
+
+### Performance
+
+* `scandir` returns `Stat`s relative to its directory. ([#19246](https://github.com/pantsbuild/pants/pull/19246))
+
 ## 2.18.0.dev1 (Jun 02, 2023)
 
 ### New Features
```

This also moves it into the new `pants_release` root, adds some basic
tests, and has it fetch the relevant branch directly, rather than
prompting the user to do so. It also pulls out a `git.py` support module
with helpers for exec-ing `git` as an external process.

The commits are individually reviewable.

This is a step towards automating more of the "start release" process,
per https://github.com/pantsbuild/pants/discussions/19279. After this
core refactoring of the functionality, I think the next step is to
combine it with the CONTRIBUTORS update and version bumping, and then
after that string it all together on CI so that starting a release is
fully automated.