A few fixes for the auto-cherry-pick script (#19307)
A few fixes:
- Stop using `if: always()`. As I'm understanding it, this will force
the job to run, even if prerequisites didn't run (I'm thinking this is
for, like critical teardown jobs/steps). Instead use `if: success() ||
failure()`, which is suggested in the docs. As I understand there are
(at least?) 4 states: `success`, `failure`, `skipped`, and `cancelled`.
- Used upstream `github-script` since we no longer talk to the web
- This fixes https://github.com/pantsbuild/pants/issues/19305 (I think)
- Pass a dummy PAT
- Add a new test for PR event
- Add a new xfail test that should be testing the bugfix, but
unfortunately `act` has a bug

So, can't test that this fixes it, but the suggestion comes from GitHub
directly:
https://docs.github.com/en/actions/learn-github-actions/expressions#always