Terraform: Add `experimental-deploy` and wire-in dependency inference (#19185)
This was actually some prework for lockfiles. But Terraform has some
ideas about differences between a "root" module (one you would deploy)
and other modules (ones which would be included in a deployment). To
model that, I creating the `TerraformDeployment` target type. Once you
have that, might as well wire it in to the `experimental-deploy` goal
(especially since that's most of the value of Terraform).
Also it looks like we weren't fetching dependencies (on other modules)
when building up the files to run Terraform commands. Now that happens.

I feel like I'm not being efficient in this MR with the way I bundle the
(pants-inferred) dependencies together with everything, since there's
this ever-growing ball of files. LMK if we've got a better way.