pex_binary sandbox: Traversing dep graph, stop after package targets (#19155)
This builds on:
- #19272
- #19306
- #19387

This only updates the `pex_binary` rule to use the new
`TraverseIfNotPackageTarget` predicate when requesting the source files
that should be included in this pex. Wheels, other pex_binaries, and
other package targets are not handled via this code path since those are
not python sources that get included in the pex based on the
`include_sources` flag.

Fixes #15855

This fixes #15855 because anything in a `python_distribution` does not
need to be included as a source file in the `pex_binary`. wheels get
included via the `LocalDists` rules. In some cases, we might still get
more sources in the pex than intended. This might happen if a dependency
is inferred between sources, bypassing the logic that looks for wheels
that own the relevant things. In any case, this is PR provides an
improvement and resolves the sample error in #15855.

Related:
- #18254
- #17368
- #15082

---------

Co-authored-by: Joshua Cannon <joshdcannon@gmail.com>