Add `gh` as a Pants tool (#19481)
This change adds `gh` as a runnable tool (i.e. `pants run
3rdparty/tools/gh -- <args>`), which is necessary for a few situations
in our GitHub Actions where we want to run `gh` in a container, and
we've already checked out the repo (we COULD download and install `gh`
on PATH, but this is simpler and is yummy dogfood)

As a result, `act` has been moved to be a neighbor of `gh` under
`3rdparty/tools`.