Link to use_rust_parser documentation for convenience (Cherry-pick of #19491) (#19492)
This adds a link the rendered web docs for
`[python-infer].use_rust_parser` to make it easier for users to action
the (good) suggestion of "Read the help for
[python-infer].use_rust_parser".

As it is, someone less familiar with pants might take a while to work
out how to do that e.g. a naive thing of `pants help
'[python-infer].use_rust_parser'` or `pants help-advanced
'[python-infer].use_rust_parser'` doesn't show it.