Automate GitHub Release notes (#19531)
Do the easy thing of just stripping the release notes out of the
relevant `notes/` file and use that to populate the GitHub Release
notes.

The addition of `get_release_notes` would need to be cherry-picked, but
not the Action yml. I'll do this manually to 2.16 and 2.17 branches.
(Alternatively we could try using the `run` step with the `python`
shell, but I haven't had great luck with that. Or we could try porting
the logic to Bash)